Remove AoeForm directory
