--- frontend/src/app/testQueue/AoEForm/AoEForm.scss
@@ -1,32 +1,0 @@
-div.usa-form-group--error {
-  .usa-form-group--error {
-    border-left-width: 0;
-    padding-left: 0;
-
-    @media screen and (min-width: 1024px) {
-      border-left-width: 0.25rem;
-      padding-left: 1rem;
-    }
-  }
-}
-
-.prime-checkboxes {
-  .usa-error-message {
-    + .usa-form-group--error {
-      margin-top: 0;
-    }
-  }
-}
-
-// Checked label description
-.usa-radio__input + .usa-radio__label {
-  .radio__label-description--checked {
-    display: none;
-  }
-}
-
-.usa-radio__input:checked + .usa-radio__label {
-  .radio__label-description--checked {
-    display: block;
-  }
-}
